Added Authentication API with testcases

+++ api/src/main/java/com/lightningcd/api/service/AuthenticationServiceImpl.java
@@ -1,13 +1,15 @@
 package com.lightningcd.api.service;
 
 
-import com.lightningcd.api.exception.UserAlreadyExistException;
+import com.lightningcd.api.exception.LoginFailedException;
 import com.lightningcd.api.exception.UserNotFoundException;
 import com.lightningcd.api.model.Authentication;
 import com.lightningcd.api.repository.AuthenticationRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
+@Service
 public class AuthenticationServiceImpl implements AuthenticationService {
 
     private AuthenticationRepository authenticationRepository;
@@ -28,40 +30,57 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     }
 
     @Override
-    public Authentication get(String username) {
-        return authenticationRepository.findByUserName(username);
-    }
-
-    @Override
-    public String create(String userName, String password) throws UserAlreadyExistException {
-        Authentication authentication = new Authentication(userName, password);
-        authenticationRepository.save(authentication);
-        return authentication.getUsername();
-    }
-
-    @Override
-    public String update(String userName, String password) throws UserNotFoundException {
-        Authentication authentication = authenticationRepository.findByUserName(userName);
-        authentication.setPassword(password);
-        authentication.setUsername(userName);
-        authenticationRepository.save(authentication);
-        return authentication.getUsername();
-    }
-
-    @Override
-    public String delete(String userName) throws UserNotFoundException {
-        Authentication authentication = authenticationRepository.findByUserName(userName);
-        authenticationRepository.delete(authentication);
-        return authentication.getUsername();
-    }
-
-    @Override
-    public String authenticateUser(String username, String password) throws UserNotFoundException {
-        Authentication authentication = authenticationRepository.findByUserName(username);
-        if (authentication.getPassword().equals(password)) {
-            return "successful";
+    public Authentication get(String username) throws UserNotFoundException {
+        Authentication authentication = authenticationRepository.findByUsername(username);
+        if (null != authentication) {
+            return authentication;
         } else {
-            return "unsuccessful";
+            throw new UserNotFoundException("User does not exist");
+        }
+    }
+
+    @Override
+    public Authentication create(String username, String password) {
+        Authentication authentication = new Authentication(username, password);
+        authenticationRepository.save(authentication);
+        return authentication;
+    }
+
+    @Override
+    public Authentication update(String username, String password) throws UserNotFoundException {
+        Authentication authentication = authenticationRepository.findByUsername(username);
+        if (null != authentication) {
+            authentication.setPassword(password);
+            authentication.setUsername(username);
+            authenticationRepository.save(authentication);
+            return authentication;
+        } else {
+            throw new UserNotFoundException("User does not exist");
+        }
+    }
+
+    @Override
+    public String delete(String username) throws UserNotFoundException {
+        Authentication authentication = authenticationRepository.findByUsername(username);
+        if (null != authentication) {
+            authenticationRepository.delete(authentication);
+            return authentication.getUsername();
+        } else {
+            throw new UserNotFoundException("User does not exist");
+        }
+    }
+
+    @Override
+    public Authentication authenticateUser(String username, String password) throws UserNotFoundException, LoginFailedException {
+        Authentication authentication = authenticationRepository.findByUsername(username);
+        if (null != authentication) {
+            if (authentication.getPassword().equals(password)) {
+                return authentication;
+            } else {
+                throw new LoginFailedException("Invalid Username or Password");
+            }
+        } else {
+            throw new UserNotFoundException("User does not exist");
         }
     }
 }
